Integration tests for mssql metadata extraction
