+++ src/integrationTest/resources/issues/siardsuite115/mssql-varchar-types.sql
